add sorting and verbose options for collectBuildInfo.sh

+++ product/collectBuildInfo.sh
@@ -8,25 +8,14 @@
 # SPDX-License-Identifier: EPL-2.0
 #
 
-# script to collect create/start/finish data on OSBS tasks. Pass in param: taskid
-# requires: brew/koji rpm to be installed
-
-# 	 Store data as yaml, with option to convert to csv
-# 	  * container that was built (incl. tag)
-# 	  	  $➔ curl -sSLo- https://download.eng.bos.redhat.com/brewroot/work/tasks/4137/48564137/x86_64.log | tail -2 | sed -r -e "s@.+containers/#/(.+)\"@\1@" | head -1
-# 	  * dates
-
-# --------
-# $➔ brew taskinfo 48564137 | grep -E "Created|Started|Finished"
-# Created: Tue Oct 25 12:39:21 2022
-# Started: Tue Oct 25 15:40:08 2022
-# Finished: Tue Oct 25 15:45:50 2022
-# --------
+# script to collect create/start/finish data on OSBS tasks. 
 
 TASK_ID=""    # required, task id from brew/osbs
 OUTPUT_YML="" # required, path and filename to update
 OUTPUT_CSV="" # optional, also write to CSV file
 APPEND=0      # by default, create a new yml / csv file
+SORT=1        # by default, sort by taskId descending
+VERBOSE=0     # by default, be quiet
 
 usage () {
 	echo "
@@ -34,19 +23,21 @@ Usage:
   $0 [-t TASK_ID | -b BUILD_ID] [OPTIONS]
 
 Options:
-  -t              Collect data for a given Task ID
-  -b              Collect data for a given Build ID
-  -f              Optionally, write to /path/to/output.yaml
-  --csv           Optionally, also write to /path/to/output.csv
-  --append        When writing to .yaml file (and .csv file), append instead of overwriting
+  -t               Collect data for a given Task ID
+  -b               Collect data for a given Build ID
+  -f               Optionally, write to /path/to/output.yaml
+  --csv            Optionally, also write to /path/to/output.csv
+  --append         When writing to .yaml file (and .csv file), append instead of overwriting
+  --unsorted       Don't sort chronologically by taskId descending
+  -v, --verbose    Verbose output: include additional information
+  -h, --help       Show this help
 
 Example - collect metadata for the current builds in 
 https://github.com/redhat-developer/devspaces/blob/devspaces-3-rhel-8/dependencies/LATEST_IMAGES_COMMITS:
 
-    for d in $(cat /path/to/LATEST_IMAGES_COMMITS | grep Build | sed -r -e "s@.+buildID=@@"); do \
-      $0 -b $d --append -f /tmp/collectBuildInfo.yml --csv /tmp/collectBuildInfo.csv ; \
+    for d in \$(cat /path/to/LATEST_IMAGES_COMMITS | grep Build | sed -r -e "s@.+buildID=@@"); do \\
+      $0 -b \$d --append -f /tmp/collectBuildInfo.yml --csv /tmp/collectBuildInfo.csv ; \\
     done
-
 "
 }
 
@@ -57,7 +48,10 @@ while [[ "$#" -gt 0 ]]; do
     '-f') OUTPUT_YML="$2"; shift 1;;
     '--csv') OUTPUT_CSV="$2"; shift 1;;
     '--append') APPEND=1;;
-    '--help') usage; exit;; 
+    '--unsorted') SORT=0;;
+    '-v'|'--verbose') VERBOSE=1;;
+    '-h'|'--help') usage;;
+    *) echo "Unknown parameter used: $1."; usage; exit 1;;
   esac
   shift 1
 done
@@ -68,12 +62,12 @@ if [[ -z ${TASK_ID} ]] && [[ -z ${BUILD_ID} ]]; then usage; exit 1; fi
 datediff() {
     d1=$(date -d "$1" +%s)
     d2=$(date -d "$2" +%s)
-    echo $(( (d2 - d1) / 60 )) minutes
+    echo -n "$(( (d2 - d1) / 60 )) mins"
 }
 
 getTaskIDFromBuildID () {
     TASK_ID=$(brew buildinfo $BUILD_ID | grep container_koji_task_id | sed -r -e "s@.+'container_koji_task_id': ([0-9]+),.+@\1@")
-    echo $TASK_ID
+    echo -n $TASK_ID
 }
 
 getContainerFromTaskID () {
@@ -84,7 +78,7 @@ getContainerFromTaskID () {
         taskid=$1
         log="https://download.eng.bos.redhat.com/brewroot/work/tasks/${taskid:(-4)}/${taskid}/x86_64.log"
         container=$(curl -sSLko- $log | tail -2 | sed -r -e "s@.+containers/#/(.+)\"@\1@" -e "s@/images/@:@" -e "s@registry.access.redhat.com/@@" | head -1)
-        echo $container
+        echo -n $container
     fi
 }
 
@@ -94,7 +88,7 @@ getContainerFromBuildID () {
     buildid=$1
     container=$(brew buildinfo $buildid | grep Extra | sed -r -e "s@Extra: @@" | yq -r '.image.index.pull[]' | grep -v sha256 \
         | sed -r -e "s@registry-proxy.engineering.redhat.com/rh-osbs/@@")
-    echo $container
+    echo -n $container
 }
 
 getADate() {
@@ -104,9 +98,17 @@ getADate() {
 }
 
 if [[ $BUILD_ID ]]; then
-    echo -n "For Build ID = $BUILD_ID: "; # getContainerFromBuildID $BUILD_ID
+    if [[ $VERBOSE -eq 1 ]]; then
+        echo -n "For Build ID = $BUILD_ID: "; # getContainerFromBuildID $BUILD_ID
+    fi
     TASK_ID=$(getTaskIDFromBuildID $BUILD_ID)
-    if [[ $TASK_ID ]]; then echo "Task ID = $TASK_ID"; else echo "[ERROR] could not compute Task ID for this Build ID"; exit 2; fi
+    if [[ $TASK_ID ]]; then 
+        if [[ $VERBOSE -eq 1 ]]; then
+            echo "Task ID = $TASK_ID"
+        fi
+    else 
+        echo "[ERROR] could not compute Task ID for this Build ID"; exit 2
+    fi
 fi
 echo -n "For Task ID = $TASK_ID: "; getContainerFromTaskID $TASK_ID
 dates="$(brew taskinfo $TASK_ID | grep -E "Created|Started|Finished")"
@@ -127,16 +129,51 @@ yaml="- name: ${container%:*}
   timeBuild: ${timeBuild}
   timeTotal: ${timeTotal}
 "
-
-echo "$yaml"
-if [[ $OUTPUT_YML ]]; then
-    if [[ $APPEND -eq 1 ]]; then
-        echo "$yaml" >> $OUTPUT_YML
-    else
-        echo "$yaml" > $OUTPUT_YML
+if [[ $VERBOSE -eq 1 ]]; then
+    echo;echo "$yaml"
+fi
+if [[ -f $OUTPUT_YML ]]; then touch $OUTPUT_YML; fi
+if [[ $(grep "  taskId: $TASK_ID" $OUTPUT_YML) ]]; then
+    if [[ $VERBOSE -eq 1 ]]; then
+        echo "[INFO] Skip: taskId $TASK_ID already in $OUTPUT_YML"; echo
+    else 
+        echo " - skipped"
     fi
-    if [[ $OUTPUT_CSV ]]; then 
-        cat $OUTPUT_YML | yq -r '.[]|(keys_unsorted)|@csv' | uniq > $OUTPUT_CSV
-        yq -r '.[]|flatten|@csv' $OUTPUT_YML >> $OUTPUT_CSV
+else
+    if [[ $OUTPUT_YML ]]; then
+        if [[ $VERBOSE -eq 0 ]]; then echo; fi
+
+        if [[ $APPEND -eq 1 ]]; then
+            echo "$yaml" >> $OUTPUT_YML
+        else
+            echo "$yaml" > $OUTPUT_YML
+        fi
+
+        # sorting removes spaces between yaml entries
+        if [[ $SORT -eq 1 ]]; then
+            # sort yaml by taskId (most recent tasks at the end of the file)
+            yq -Y -i '.|=sort_by(.taskId)' $OUTPUT_YML
+        fi
+
+        if [[ $OUTPUT_CSV ]]; then 
+            if [[ $SORT -eq 1 ]]; then
+                # always replace csv file with fresh content from the yaml file
+                cat $OUTPUT_YML | yq -r '.[]|(keys_unsorted)|@csv' | uniq > $OUTPUT_CSV
+                cat $OUTPUT_YML | yq -r '.[]|flatten|@csv' >> $OUTPUT_CSV
+            else
+                # add header if needed (file doesn't exist or is empty)
+                if [[ ! -f $OUTPUT_CSV ]] || [[ ! -s $OUTPUT_CSV ]]; then
+                    cat $OUTPUT_YML | yq -r '.[]|(keys_unsorted)|@csv' | uniq > $OUTPUT_CSV
+                fi
+                # add the new yaml as a line of csv
+                echo "$yaml" | yq -r '.[]|flatten|@csv' >> $OUTPUT_CSV
+            fi
+        fi
+
+        if [[ $VERBOSE -eq 1 ]]; then
+            echo "[INFO] Wrote info to $OUTPUT_YML"
+            if [[ $OUTPUT_CSV ]]; then echo "[INFO] Wrote info to $OUTPUT_CSV"; fi
+            echo
+        fi
     fi
 fi
